Enable PNG compression back
A while back, we ran the images through ImageOptim and reduced their size.
Since you have to remember to do this for every new image and we didn't, the savings in size are not huge.

Sizes of release .ipa file
* Current revision: 3.3 MB
* After enabling png compression: 3.5 MB
* Running ImageOptim on the new images: 2.7 MB

So at most, we'd be saving 800K in download size

On the other hand, Xcode's PNG compression is more optimized for loading images on iOS, so I think we can trade those 800K for more speed in image loading.

See:

* http://bjango.com/articles/pngcompression/
* http://imageoptim.com/xcode.html

git-svn-id: http://iphone.svn.wordpress.org/trunk@3186 7d847f7c-ffe1-4a24-85be-b788d016d96a