Fixes #1333.  If the user's Jetpack credentials were invalid there was no way to update them.
Allows the user to be prompted to update their credentials if Jetpack fails to auth.
Authentication is confirmed by validating that the login cookie is set rather than relying on .com to return a 401 error for invalid auth.
Removes the code that would try attempt to pre-auth stats. It wasn't contributing to #1333 since it wasn't in the 3.1 release but seemed unnecessarily complex.
Listens for sign out notifications to clear the lastLoggedIn user name flag.





git-svn-id: http://iphone.svn.wordpress.org/trunk@3364 7d847f7c-ffe1-4a24-85be-b788d016d96a