Remove old notes
We trigger the note pruning on pull to refresh, or when the
notifications list disappears.

To avoid making notes disappear in front of the user, we'll keep notes
up to the selected one or whatever is visible on the table.

Otherwise, we keep the last 20, since we fetch batches of 20 on sync.
This helps when there are more than 20 notifications when syncing, as
all the old ones would drop. If we kept more, there would be a "gap" of
notifications not accessible between the old ones and the notes just
synced.

fixes #1519

git-svn-id: http://iphone.svn.wordpress.org/trunk@3890 7d847f7c-ffe1-4a24-85be-b788d016d96a